+++ beacon-chain/p2p/BUILD.bazel
@@ -96,6 +96,7 @@ go_test(
     srcs = [
         "addr_factory_test.go",
         "broadcaster_test.go",
+        "connection_gater_test.go",
         "dial_relay_node_test.go",
         "discovery_test.go",
         "fork_test.go",
