+++ beacon-chain/p2p/options_test.go
@@ -2,18 +2,13 @@ package p2p
 
 import (
 	"bytes"
-	"context"
 	"crypto/rand"
 	"encoding/hex"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"testing"
 
-	"github.com/libp2p/go-libp2p"
 	"github.com/libp2p/go-libp2p-core/crypto"
-	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/multiformats/go-multiaddr"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 )
 
@@ -61,56 +56,5 @@ func TestPrivateKeyLoading(t *testing.T) {
 	}
 	if !bytes.Equal(newRaw, rawBytes) {
 		t.Errorf("Private keys do not match got %#x but wanted %#x", rawBytes, newRaw)
-	}
-}
-
-func TestPeerDenyList(t *testing.T) {
-	// create host with deny list
-	ipAddr, pkey := createAddrAndPrivKey(t)
-	ipAddr2, pkey2 := createAddrAndPrivKey(t)
-
-	mask := ipAddr2.DefaultMask()
-	ones, _ := mask.Size()
-	maskedIP := ipAddr2.Mask(mask)
-	cidr := maskedIP.String() + fmt.Sprintf("/%d", ones)
-
-	listen, err := multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/%s/tcp/%d", ipAddr, 2000))
-	if err != nil {
-		t.Fatalf("Failed to p2p listen: %v", err)
-	}
-	h1, err := libp2p.New(context.Background(), []libp2p.Option{privKeyOption(pkey), libp2p.ListenAddrs(listen), denyListSubnets([]string{cidr})}...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		err := h1.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-	}()
-
-	// create alternate host
-	listen, err = multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/%s/tcp/%d", ipAddr2, 3000))
-	if err != nil {
-		t.Fatalf("Failed to p2p listen: %v", err)
-	}
-	h2, err := libp2p.New(context.Background(), []libp2p.Option{privKeyOption(pkey2), libp2p.ListenAddrs(listen)}...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		err := h2.Close()
-		if err != nil {
-			t.Fatal(err)
-		}
-	}()
-	multiAddress, err := multiaddr.NewMultiaddr(fmt.Sprintf("/ip4/%s/tcp/%d/p2p/%s", ipAddr2, 3000, h2.ID()))
-	addrInfo, err := peer.AddrInfoFromP2pAddr(multiAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = h1.Connect(context.Background(), *addrInfo)
-	if err == nil {
-		t.Error("Wanted connection to fail with deny list")
 	}
 }
